Started a major refactor
Started work on actually generating bytecode output - which lead to this refactor. There's still mroe to do, but this is a good start.

I'm trying to figure out how to map the stuff from Crafting Interpreters to Rust ... while learning Rust; it's fairly tricky. Expect more big changes like this for the next few commits ...