Working through Crafting Interpreters
Doing my best to translate the C code to Rust as I go. This commit cover chapters 14 & 15.

I do deviate from the book in a few places - we'll see how that goes 😅

Also, there code that's commented out - because I'm expecting I'll use it, or some version of it, later. Again... we'll see how that goes...