Removed all line-tracking logic & tests
This feels very cumbersome; I feel like there must be a better way to handle this.

Once I have a better grasp of all this, I'll revisit error tracking/reporting - which, it feels like, is the only case where this will be useful.

I'm thinking that line/column data should live in its own data structure... just not sure of the details yet.